Avoid subtraction overflow in SplitByLogdChunks().
Passing a _tag_ (as opposed to a message) too long to fit in a single
log message is obviously a mistake. But it happened in the wild, and
caused Java code to die with a ubsan trap.

There's an argument that the point at which we call a tag "unreasonable"
should be no more than _half_ the buffer, but this corresponds more
closely with the lower-level liblog implementation which will happily
let you use every last byte on your tag.

Bug: https://issuetracker.google.com/400886840
Test: new test
Change-Id: I0a0980dcc75e87fc72cca9787aade5dfd554972d