Fix enabling of messages using log.tag.<tag> property
Summary:
aosp/1211227 changed the loggability checks used by the LOG/LOG_STREAM
macros to use liblog's __android_log_is_loggable function but enabling
logs using the per-tag log.tag.<tag> property doesn't work because the
log tag used to check loggability in the LogMessage destructor is not
the log tag with which the message was logged but rather that of
logging.cpp which is not defined so liblog ends up checking against
the log.tag.<tag> property for the default tag. Fix this by using the
log message's log tag when performing the loggability check in the
LogMessage destructor. Also modify the LOG_xxx_enabled device tests
to verify that the per-tag log.tag.<tag> property correctly enables
logging.

Test: atest libbase_test on aosp_cheetah-bp1a-eng
Test: atest --host libbase_test
Test: Removing the usage of the new macro and verifying the tests fail.
Change-Id: I9cf9a299af7f038bb452b5fae17e384b59d16e56