Remove unneeded use of ASSERT_NO_FATAL_FAILURE.
The ASSERT_NO_FATAL_FAILURE is used to allow a failure in a function
to stop the code afterwards from executing. It doesn't make sense
to add it to a single call, or add it to the last function called
in a test.

Remove all of the unnecessary uses of the macro.

Test: All unit tests pass.
Change-Id: I52db7f1789a9afbe8fddb35a0c8e59014b00724b